+   Encountered an irritating issue with packet decoding & resolved it after spending a day on it. 		-   Finished ISAAC decryption with decoding 	+   Finally finished player file synchronization over multiple worlds 		-   Merged lobby server & download server into one server, with different handshake handling. 	+   Started PVM System 		-   Loaded 971 dumped npc combat definitions 			+   These include 1000+ ids of npcs with different bonuses, attack animations, etc 			+   As well as charm drops 			+   As well as drop tables 		-   Wrote npc drop table system & implemented it [with charms] 		-   Wrote npc combat system and implemented it with a test class for npcs that don't have custom swings 			+   Man, guard, woman, hellhound, etc.... 		-   Wrote npc defense bonus calculation for pvm [so realistic hits are done when you're fighting npcs] 	+   Added wilderness area checking to player appearance update block [126 + 12/138 cb lvl]

+++ src/org/redrune/game/node/entity/player/data/PlayerRight.java
@@ -35,10 +35,10 @@ public enum PlayerRight {
 	 * The rights the player has in the client
 	 */
 	@Getter
-	private final int clientRight;
+	private final byte clientRight;
 	
 	PlayerRight(int clientRight) {
-		this.clientRight = clientRight;
+		this.clientRight = (byte) clientRight;
 	}
 	
 	PlayerRight() {
